fix: fix codegen/macro alignment
